% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{trace.plot}
\alias{trace.plot}
\title{Generate a trace plot from an mcmc run}
\usage{
trace.plot(mcmc, var.names = mcmc@control@var.names,
  chains = 1:mcmc@n.chains, exact.var.names = F)
}
\arguments{
\item{mcmc}{A \link{mcmc.sim} object}

\item{var.names, exact.var.names}{var.names is the names of variables to include or a set of integer indices. If exact.var.names is TRUE, matches the given strings exactly. If exact.var.names is FALSE, matches any variable name that begins with any element of var.names, where the * character is treated as a wild card}

\item{chains}{Which chains from the MCMC to use}
}
\description{
Generate a trace plot from an mcmc run
}
